+++ Mermaid_General_Pic-Web/src/views/Home.vue
@@ -1,831 +1,39 @@
 <template>
   <div class="home">
-    <!-- 顶部导航栏 -->
-    <el-header class="header">
-      <div class="nav-container">
-        <div class="nav-left">
-          <h2 class="logo">{{ t('editor.title') }}</h2>
-        </div>
-        <div class="nav-right">
-          <el-menu
-            mode="horizontal"
-            :router="true"
-            :default-active="$route.path"
-            class="nav-menu"
-          >
-            <el-menu-item index="/">
-              <el-icon><HomeFilled /></el-icon>
-              {{ t('nav.home') }}
-            </el-menu-item>
-            <el-menu-item index="/history">
-              <el-icon><Timer /></el-icon>
-              {{ t('history.title') }}
-            </el-menu-item>
-          </el-menu>
-        </div>
-      </div>
-    </el-header>
-
+    <NavBar />
     <el-container>
       <el-main>
         <!-- AI 对话栏 -->
         <el-row :gutter="20" class="mb-4">
           <el-col :span="24">
-            <el-card class="ai-card" shadow="hover">
-              <div class="ai-settings">
-                <el-select
-                  v-model="aiStore.settings.selectedModel"
-                  @change="handleModelChange"
-                  class="model-select"
-                  :placeholder="t('ai.model')"
-                >
-                  <el-option
-                    v-for="(config, model) in aiStore.settings.modelConfigs"
-                    :key="model"
-                    :label="config.name"
-                    :value="model"
-                  />
-                </el-select>
-                <el-select
-                  v-if="isKimiModel"
-                  v-model="kimiModel"
-                  @change="handleKimiModelChange"
-                  class="model-version-select"
-                  :placeholder="t('ai.kimi_model_placeholder')"
-                >
-                  <el-option
-                    v-for="(label, value) in kimiModels"
-                    :key="value"
-                    :label="label"
-                    :value="value"
-                  />
-                </el-select>
-                <el-input
-                  v-model="currentApiKey"
-                  type="password"
-                  :placeholder="t('ai.api_key_placeholder')"
-                  show-password
-                  @change="handleApiKeyChange"
-                  class="api-key-input"
-                />
-              </div>
-              <div class="ai-input-container">
-                <el-input
-                  v-model="aiPrompt"
-                  :placeholder="t('ai.placeholder')"
-                  :disabled="isGenerating"
-                  @keyup.enter="handleGenerate"
-                >
-                  <template #prepend>
-                    <el-button
-                      type="primary"
-                      plain
-                      @click="handleNewConversation"
-                      :disabled="isGenerating"
-                    >
-                      {{ t('ai.new_conversation') }}
-                    </el-button>
-                  </template>
-                  <template #append>
-                    <el-button
-                      type="primary"
-                      :loading="isGenerating"
-                      @click="handleGenerate"
-                    >
-                      {{ t('ai.generate') }}
-                    </el-button>
-                  </template>
-                </el-input>
-              </div>
-
-              <!-- 对话历史记录 -->
-              <div v-if="chatHistory.length > 0" class="chat-history">
-                <div v-for="(message, index) in chatHistory" :key="index" class="chat-message">
-                  <div class="message-header">
-                    <el-tag size="small" :type="message.role === 'user' ? 'primary' : 'success'">
-                      {{ message.role === 'user' ? t('ai.user') : t('ai.assistant') }}
-                    </el-tag>
-                    <span class="message-time">{{ formatTime(message.timestamp) }}</span>
-                  </div>
-                  <div class="message-content" :class="{ 'user-message': message.role === 'user' }">
-                    {{ message.content }}
-                  </div>
-                  <div v-if="message.code" class="message-code">
-                    <pre>{{ message.code }}</pre>
-                  </div>
-                </div>
-              </div>
-            </el-card>
+            <AiChat ref="aiChat" />
           </el-col>
         </el-row>
 
-        <el-row :gutter="20">
-          <!-- 左侧编辑器 -->
-          <el-col :span="12">
-            <el-card class="editor-card" shadow="hover">
-              <template #header>
-                <div class="card-header">
-                  <div class="header-title">
-                    <el-icon><Edit /></el-icon>
-                    <span>{{ t('editor.title') }}</span>
-                  </div>
-                  <div class="header-actions">
-                    <el-upload
-                      class="upload-btn"
-                      :show-file-list="false"
-                      accept=".txt,.mmd"
-                      :before-upload="handleUpload"
-                    >
-                      <el-button type="primary" size="small">
-                        <el-icon><Upload /></el-icon>
-                        {{ t('editor.upload') }}
-                      </el-button>
-                    </el-upload>
-                  </div>
-                </div>
-              </template>
-              
-              <div class="editor-container">
-                <div class="editor-main">
-                  <el-input
-                    v-model="store.code"
-                    type="textarea"
-                    :rows="25"
-                    :placeholder="t('editor.placeholder')"
-                    @input="handleCodeChange"
-                    resize="none"
-                    spellcheck="false"
-                  />
-                </div>
-
-                <div class="editor-footer">
-                  <el-button
-                    type="primary"
-                    :loading="store.isLoading"
-                    @click="handleConvert"
-                    size="small"
-                  >
-                    <el-icon><Refresh /></el-icon>
-                    {{ t('editor.convert') }}
-                  </el-button>
-                </div>
-              </div>
-
-              <el-alert
-                v-if="store.error"
-                :title="store.error"
-                type="error"
-                show-icon
-                class="error-alert"
-              />
-            </el-card>
-          </el-col>
-
-          <!-- 右侧预览 -->
-          <el-col :span="12">
-            <el-card class="preview-card" shadow="hover">
-              <template #header>
-                <div class="card-header">
-                  <div class="header-title">
-                    <el-icon><View /></el-icon>
-                    <span>{{ t('preview.title') }}</span>
-                  </div>
-                  <div class="preview-controls">
-                    <div v-if="previewUrl">
-                      <el-input-number
-                        v-model="zoomPercent"
-                        :min="10"
-                        :max="300"
-                        :step="10"
-                        size="small"
-                        style="width: 110px"
-                        @change="handleZoomChange"
-                        controls-position="right"
-                      >
-                        <template #suffix>%</template>
-                      </el-input-number>
-                    </div>
-                    <DownloadDialog
-                      v-if="previewUrl"
-                      :code="store.code"
-                    />
-                  </div>
-                </div>
-              </template>
-              <div class="preview-container">
-                <el-image
-                  v-if="previewUrl"
-                  :src="previewUrl"
-                  fit="contain"
-                  :style="{ transform: `scale(${zoomLevel})`, transformOrigin: 'top left' }"
-                  class="preview-image"
-                  :preview-src-list="[previewUrl]"
-                  :initial-index="0"
-                  preview-teleported
-                  :preview-options="{
-                    zoom: 1.5,
-                    zoomRate: 1.2,
-                    defaultFullScreen: false,
-                    closeOnPressEscape: true,
-                    teleported: true
-                  }"
-                >
-                  <template #placeholder>
-                    <div class="image-placeholder">
-                      <el-icon><Loading /></el-icon>
-                      <span>{{ t('preview.loading') }}</span>
-                    </div>
-                  </template>
-                  <template #error>
-                    <div class="image-error">
-                      <el-icon><PictureFilled /></el-icon>
-                      <span>{{ t('preview.error') }}</span>
-                    </div>
-                  </template>
-                </el-image>
-                <div v-else class="empty-preview">
-                  <el-icon><Picture /></el-icon>
-                  <span>{{ t('preview.empty') }}</span>
-                </div>
-              </div>
-            </el-card>
-          </el-col>
-        </el-row>
+        <!-- Mermaid 编辑器和预览 -->
+        <MermaidEditor ref="mermaidEditor" />
       </el-main>
     </el-container>
   </div>
 </template>
 
 <script setup lang="ts">
-import { ref, onMounted, computed } from 'vue';
-import {
-  Edit,
-  Upload,
-  Picture,
-  View,
-  Loading,
-  PictureFilled,
-  Refresh,
-  Timer,
-  HomeFilled
-} from '@element-plus/icons-vue';
-import { useMermaidStore } from '../stores/mermaid';
-import { useAiStore } from '../stores/ai';
-import DownloadDialog from '../components/DownloadDialog.vue';
-import { useI18n } from 'vue-i18n';
-import { api } from '../services/api';
-import { ElMessage } from 'element-plus';
-import { AI_MODELS, KIMI_MODELS, type AiModel, type ChatMessage } from '../types';
-import { useRouter } from 'vue-router';
+import { ref } from 'vue';
+import NavBar from '../components/NavBar.vue';
+import AiChat from '../components/AiChat.vue';
+import MermaidEditor from '../components/MermaidEditor.vue';
 
-const { t } = useI18n();
-const store = useMermaidStore();
-const aiStore = useAiStore();
-const previewUrl = ref<string>('');
-const zoomLevel = ref(1);
-const zoomPercent = ref(100);
-const aiPrompt = ref('');
-const isGenerating = ref(false);
-const conversationId = ref<string>('');
-const chatHistory = ref<ChatMessage[]>([]);
-const currentApiKey = ref(aiStore.getCurrentApiKey());
-const kimiModel = ref(aiStore.getCurrentKimiModel());
-const kimiModels = KIMI_MODELS;
-const isKimiModel = computed(() => aiStore.settings.selectedModel === AI_MODELS.KIMI);
-const router = useRouter();
-
-// 处理模型变更
-const handleModelChange = (model: AiModel) => {
-  aiStore.updateSelectedModel(model);
-  currentApiKey.value = aiStore.getCurrentApiKey();
-};
-
-// 处理API密钥变更
-const handleApiKeyChange = (apiKey: string) => {
-  aiStore.updateApiKey(aiStore.settings.selectedModel, apiKey);
-};
-
-// 处理 Kimi 模型变更
-const handleKimiModelChange = (model: string) => {
-  aiStore.updateKimiModel(model);
-};
-
-// 初始化
-onMounted(async () => {
-  await store.initializeStore();
-});
-
-// 处理代码变更
-let timer: number | null = null;
-const handleCodeChange = () => {
-  if (timer) {
-    clearTimeout(timer);
-  }
-  timer = window.setTimeout(async () => {
-    if (store.code) {
-      await handleConvert(false);
-    }
-  }, 1000);
-};
-
-// 处理转换
-const handleConvert = async (addToHistory: boolean = true) => {
-  try {
-    // 预览时使用更高的 DPI 以获得更清晰的图像
-    const previewDpi = Math.max(300, window.devicePixelRatio * 300);
-    const blob = await api.convert(store.code, store.format, previewDpi, store.theme, store.background);
-    if (blob) {
-      // 如果之前有预览 URL，先释放它
-      if (previewUrl.value) {
-        URL.revokeObjectURL(previewUrl.value);
-      }
-
-      // 为预览创建一个 URL
-      previewUrl.value = URL.createObjectURL(blob.slice(0));
-      zoomLevel.value = 1;
-
-      // 只有在需要添加到历史记录时才创建新的 URL 和添加记录
-      if (addToHistory) {
-        // 为历史记录创建一个独立的 URL（使用 blob 的副本）
-        const historyUrl = URL.createObjectURL(blob.slice(0));
-        
-        // 添加到历史记录
-        store.history.unshift({
-          code: store.code,
-          format: store.format,
-          dpi: previewDpi,
-          theme: store.theme,
-          background: store.background,
-          timestamp: Date.now(),
-          url: historyUrl
-        });
-      }
-    }
-  } catch (err) {
-    console.error('Error converting:', err);
-  }
-};
-
-// 处理上传
-const handleUpload = async (file: File) => {
-  try {
-    const reader = new FileReader();
-    reader.onload = async (e) => {
-      if (e.target?.result) {
-        store.code = e.target.result.toString();
-        await handleConvert();
-      }
-    };
-    reader.readAsText(file);
-  } catch (err) {
-    console.error('Error uploading:', err);
-  }
-  return false;
-};
-
-// 处理缩放
-const handleZoomChange = (value: number) => {
-  zoomLevel.value = value / 100;
-};
-
-// 格式化时间
-const formatTime = (timestamp: number) => {
-  return new Date(timestamp).toLocaleTimeString();
-};
-
-// 处理新建对话
-const handleNewConversation = () => {
-  conversationId.value = '';
-  chatHistory.value = [];
-  ElMessage.success(t('ai.new_conversation_success'));
-};
-
-// 处理AI生成
-const handleGenerate = async () => {
-  if (!aiPrompt.value.trim()) {
-    ElMessage.warning(t('ai.prompt_required'));
-    return;
-  }
-
-  const apiKey = aiStore.getCurrentApiKey();
-  if (!apiKey) {
-    ElMessage.warning(t('ai.api_key_required'));
-    return;
-  }
-
-  isGenerating.value = true;
-
-  try {
-    // 添加用户消息到历史记录
-    chatHistory.value.push({
-      role: 'user',
-      content: aiPrompt.value,
-      timestamp: Date.now()
-    });
-
-    const response = await api.generateFromAi(aiPrompt.value, {
-      model: aiStore.settings.selectedModel as AiModel,
-      conversationId: conversationId.value,
-      apiKey: apiKey,
-      kimiModel: isKimiModel.value ? aiStore.getCurrentKimiModel() : undefined
-    });
-    
-    // 添加助手回复到历史记录
-    chatHistory.value.push({
-      role: 'assistant',
-      content: t('ai.generation_complete'),
-      code: response.code,
-      timestamp: Date.now()
-    });
-
-    store.code = response.code;
-    conversationId.value = response.conversationId;
-    await handleConvert();
-    
-    ElMessage.success(t('ai.generation_success'));
-    aiPrompt.value = '';
-  } catch (error) {
-    console.error('Failed to generate diagram:', error);
-    ElMessage.error(t('ai.generation_error'));
-  } finally {
-    isGenerating.value = false;
-  }
-};
-
-// 处理查看历史
-const handleViewHistory = () => {
-  router.push('/history');
-};
+const aiChat = ref();
+const mermaidEditor = ref();
 </script>
 
 <style scoped>
 .home {
-  padding: 20px;
   background-color: var(--el-bg-color-page);
   min-height: 100vh;
 }
 
 .mb-4 {
-  margin-bottom: 16px;
-}
-
-.ai-card {
-  border-radius: 8px;
-  box-shadow: var(--el-box-shadow-light);
-}
-
-.ai-input-container {
-  display: flex;
-  gap: 12px;
-}
-
-:deep(.el-input-group__append) {
-  padding: 0;
-}
-
-:deep(.el-input-group__append button) {
-  border: none;
-  height: 100%;
-  padding: 0 20px;
-  border-radius: 0 4px 4px 0;
-}
-
-.editor-card,
-.preview-card {
-  height: 100%;
-  border-radius: 8px;
-  box-shadow: var(--el-box-shadow-light);
-}
-
-.card-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.header-title {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.editor-container {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.editor-main {
-  position: relative;
-}
-
-:deep(.el-textarea__inner) {
-  font-family: 'Fira Code', monospace;
-  font-size: 14px;
-  line-height: 1.6;
-  padding: 12px;
-  border-radius: 8px;
-  background-color: var(--el-bg-color);
-  border: 1px solid var(--el-border-color-light);
-  transition: all 0.3s ease;
-}
-
-:deep(.el-textarea__inner:hover) {
-  border-color: var(--el-border-color-hover);
-}
-
-:deep(.el-textarea__inner:focus) {
-  border-color: var(--el-color-primary);
-  box-shadow: 0 0 0 2px var(--el-color-primary-light-8);
-}
-
-.editor-footer {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.preview-container {
-  height: 600px;
-  overflow: auto;
-  background-color: var(--el-bg-color);
-  border-radius: 8px;
-  border: 1px solid var(--el-border-color-light);
-  padding: 16px;
-}
-
-.preview-image {
-  transition: transform 0.3s ease;
-}
-
-:deep(.el-image) {
-  width: auto !important;
-  display: inline-block;
-}
-
-.empty-preview,
-.image-placeholder,
-.image-error {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 12px;
-  color: var(--el-text-color-secondary);
-  font-size: 14px;
-}
-
-.preview-controls {
-  display: flex;
-  gap: 12px;
-  align-items: center;
-}
-
-:deep(.el-button-group) {
-  display: flex;
-  align-items: center;
-}
-
-:deep(.el-button-group .el-button) {
-  padding: 8px 12px;
-}
-
-.error-alert {
-  margin-top: 16px;
-}
-
-@media (max-width: 768px) {
-  .el-row {
-    flex-direction: column;
-  }
-
-  .el-col {
-    width: 100% !important;
-    margin-bottom: 20px;
-  }
-}
-
-:deep(.el-image-viewer__wrapper) {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: 2000;
-}
-
-:deep(.el-image-viewer__mask) {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  opacity: 0.7;
-  background: #000;
-}
-
-:deep(.el-image-viewer__actions) {
-  opacity: 0.9;
-  padding: 20px;
-}
-
-:deep(.el-image-viewer__actions__inner) {
-  border-radius: 20px;
-  padding: 8px 16px;
-  background-color: rgba(0, 0, 0, 0.7);
-}
-
-:deep(.el-image-viewer__btn) {
-  color: #fff;
-  opacity: 0.8;
-  font-size: 24px;
-  cursor: pointer;
-  transition: all 0.3s;
-}
-
-:deep(.el-image-viewer__btn:hover) {
-  opacity: 1;
-  transform: scale(1.1);
-}
-
-:deep(.el-image-viewer__close) {
-  top: 20px;
-  right: 20px;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background-color: rgba(0, 0, 0, 0.7);
-  font-size: 24px;
-  opacity: 0.8;
-  transition: all 0.3s;
-}
-
-:deep(.el-image-viewer__close:hover) {
-  opacity: 1;
-  transform: rotate(90deg);
-  background-color: rgba(0, 0, 0, 0.9);
-}
-
-:deep(.el-image-viewer__canvas) {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-:deep(.el-image-viewer__img) {
-  max-width: 100%;
-  max-height: 100%;
-  user-select: none;
-  -webkit-user-drag: none;
-  cursor: grab;
-}
-
-:deep(.el-image-viewer__img:active) {
-  cursor: grabbing;
-}
-
-:deep(.el-image-viewer__prev),
-:deep(.el-image-viewer__next) {
-  display: none;
-}
-
-.chat-history {
-  margin-top: 16px;
-  max-height: 400px;
-  overflow-y: auto;
-  border-top: 1px solid var(--el-border-color-light);
-  padding-top: 16px;
-}
-
-.chat-message {
-  margin-bottom: 16px;
-  padding: 12px;
-  border-radius: 8px;
-  background-color: var(--el-bg-color-page);
-}
-
-.message-header {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 8px;
-}
-
-.message-time {
-  font-size: 12px;
-  color: var(--el-text-color-secondary);
-}
-
-.message-content {
-  margin-bottom: 8px;
-  white-space: pre-wrap;
-  word-break: break-word;
-}
-
-.message-code {
-  background-color: var(--el-bg-color);
-  border-radius: 4px;
-  padding: 12px;
-}
-
-.message-code pre {
-  margin: 0;
-  white-space: pre-wrap;
-  word-break: break-word;
-  font-family: 'Fira Code', monospace;
-  font-size: 13px;
-}
-
-.user-message {
-  color: var(--el-color-primary);
-}
-
-.ai-settings {
-  display: flex;
-  gap: 12px;
-  margin-bottom: 16px;
-}
-
-.model-select {
-  width: 150px;
-}
-
-.api-key-input {
-  flex: 1;
-}
-
-.model-version-select {
-  width: 180px;
-}
-
-.header-actions {
-  display: flex;
-  gap: 12px;
-  align-items: center;
-}
-
-.header {
-  background-color: var(--el-bg-color);
-  border-bottom: 1px solid var(--el-border-color-light);
-  padding: 0;
-  position: sticky;
-  top: 0;
-  z-index: 100;
-}
-
-.nav-container {
-  max-width: 1400px;
-  margin: 0 auto;
-  height: 60px;
-  padding: 0 20px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.nav-left {
-  display: flex;
-  align-items: center;
-}
-
-.nav-right {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.logo {
-  margin: 0;
-  font-size: 20px;
-  color: var(--el-text-color-primary);
-}
-
-.nav-menu {
-  background-color: transparent;
-  border-bottom: none;
-}
-
-:deep(.el-menu--horizontal) {
-  border-bottom: none;
-}
-
-:deep(.el-menu-item) {
-  font-size: 16px;
-  height: 60px;
-  line-height: 60px;
-  padding: 0 20px;
-}
-
-:deep(.el-menu-item.is-active) {
-  font-weight: 600;
-}
-
-:deep(.el-menu-item .el-icon) {
-  margin-right: 4px;
-  font-size: 18px;
+  margin-bottom: 24px;
 }
 </style>
