+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "@vueuse/core": "^12.4.0"
+  }
+}
